fixed bigdecimal issues

+++ U1-M5-Summative-Junghoon-Yoon/JunghoonYoonU1M5Summative/src/main/java/com/company/JunghoonYoonU1M5Summative/dao/BookDaoImpl.java
@@ -124,7 +124,7 @@ public class BookDaoImpl implements BookDAO {
         book.setAuthor_id(rs.getInt("author_id"));
         book.setTitle(rs.getString("title"));
         book.setPublisher_id(rs.getInt("publisher_id"));
-        book.setPrice(rs.getDouble("price"));
+        book.setPrice(rs.getBigDecimal("price"));
 
         return book;
     }
